Fix 64-bit compile

+++ src/emu/restrack.c
@@ -189,7 +189,7 @@ static object_pool *current_pool(void)
 
 void *restrack_register_object(object_type type, void *ptr, size_t size, const char *file, int line)
 {
-if (resource_tracking_tag == 2) mame_printf_warning("restrack_register_object(%p,%d) called within reset scope by %s, line %d\n", ptr, size, file, line);
+if (resource_tracking_tag == 2) mame_printf_warning("restrack_register_object(%p,%ld) called within reset scope by %s, line %d\n", ptr, (long) size, file, line);
 	return pool_object_add_file_line(current_pool(), type, ptr, size, file, line);
 }
 
@@ -202,7 +202,7 @@ if (resource_tracking_tag == 2) mame_printf_warning("restrack_register_object(%p
 void *auto_malloc_file_line(running_machine *machine, size_t size, const char *file, int line)
 {
 	void *result = pool_malloc_file_line(current_pool(), size, file, line);
-if (resource_tracking_tag == 2) mame_printf_warning("auto_malloc(%d) called within reset scope by %s, line %d\n", size, file, line);
+if (resource_tracking_tag == 2) mame_printf_warning("auto_malloc(%ld) called within reset scope by %s, line %d\n", (long) size, file, line);
 #ifdef MAME_DEBUG
 	rand_memory(result, size);
 #endif
@@ -218,7 +218,7 @@ if (resource_tracking_tag == 2) mame_printf_warning("auto_malloc(%d) called with
 void *auto_realloc_file_line(running_machine *machine, void *ptr, size_t size, const char *file, int line)
 {
 	object_pool *pool = current_pool();
-if (resource_tracking_tag == 2) mame_printf_warning("auto_realloc(%p, %d) called within reset scope by %s, line %d\n", ptr, size, file, line);
+if (resource_tracking_tag == 2) mame_printf_warning("auto_realloc(%p, %ld) called within reset scope by %s, line %d\n", ptr, (long) size, file, line);
 	if (ptr != NULL)
 	{
 		int tag = resource_tracking_tag;
